+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 
 ## [Unreleased]
 
-- KotlinxAndroidSyntheticToViewBinding
+- `20230312` Add ViewBinding Migration action on fragment related View.
+- `20230308` Add ViewBinding Migration action on activity related View.